Clickable repository items And a refactor: move repository related methods on activity class to fragment.
