+++ app/src/main/java/org/sorz/lab/smallcloudemoji/activites/SettingsActivity.java
@@ -2,34 +2,19 @@ package org.sorz.lab.smallcloudemoji.activites;
 
 import android.app.ActionBar;
 import android.app.Activity;
-import android.app.AlertDialog;
 import android.app.Fragment;
 import android.app.FragmentManager;
-import android.content.DialogInterface;
 import android.content.res.Configuration;
 import android.os.Bundle;
 import android.view.MenuItem;
 import android.view.View;
-import android.widget.EditText;
-import android.widget.ImageButton;
-import android.widget.PopupMenu;
-import android.widget.ProgressBar;
 
 import org.sorz.lab.smallcloudemoji.R;
-import org.sorz.lab.smallcloudemoji.db.Category;
-import org.sorz.lab.smallcloudemoji.db.CategoryDao;
 import org.sorz.lab.smallcloudemoji.db.DaoSession;
 import org.sorz.lab.smallcloudemoji.db.DatabaseHelper;
 import org.sorz.lab.smallcloudemoji.db.DatabaseUpgrader;
-import org.sorz.lab.smallcloudemoji.db.EntryDao;
-import org.sorz.lab.smallcloudemoji.db.Repository;
-import org.sorz.lab.smallcloudemoji.db.RepositoryDao;
 import org.sorz.lab.smallcloudemoji.fragments.RepositoryFragment;
 import org.sorz.lab.smallcloudemoji.fragments.SettingsFragment;
-import org.sorz.lab.smallcloudemoji.tasks.DownloadXmlAsyncTask;
-
-import java.util.ArrayList;
-import java.util.List;
 
 
 public class SettingsActivity extends Activity implements
@@ -37,7 +22,6 @@ public class SettingsActivity extends Activity implements
         SettingsFragment.OnSourceManageClickListener {
     private final static String REPOSITORY_FRAGMENT_IS_SHOWING = "REPOSITORY_FRAGMENT_IS_SHOWING";
     private RepositoryFragment repositoryFragment;
-    private RepositoryDao repositoryDao;
     private DaoSession daoSession;
 
     @Override
@@ -48,7 +32,6 @@ public class SettingsActivity extends Activity implements
         // Open database.
         DatabaseHelper databaseHelper = DatabaseHelper.getInstance(this, true);
         daoSession = databaseHelper.getDaoSession();
-        repositoryDao = daoSession.getRepositoryDao();
         DatabaseUpgrader.checkAndDoUpgrade(this, daoSession);
 
         setContentView(R.layout.activity_settings);
@@ -95,7 +78,7 @@ public class SettingsActivity extends Activity implements
         super.onConfigurationChanged(newConfig);
     }
 
-        @Override
+    @Override
     public void onSourceManageClick() {
         FragmentManager fragmentManager = getFragmentManager();
         Fragment settingsFragment = fragmentManager.findFragmentById(R.id.settings_frag);
@@ -108,187 +91,20 @@ public class SettingsActivity extends Activity implements
                 .commit();
     }
 
-    private Repository getRepositoryFromView(View view) {
-        return (Repository) ((View) view.getParent()).getTag();
-    }
-
     public void hideRepository(View view) {
-        ImageButton button = (ImageButton) view;
-        Repository repository = getRepositoryFromView(view);
-        repository.setHidden(!repository.getHidden());
-        repository.update();
-        if (repository.getHidden())
-            button.setBackgroundResource(R.drawable.ic_eye_slash);
-        else
-            button.setBackgroundResource(R.drawable.ic_eye_normal);
+        repositoryFragment.hideRepository(view);
     }
 
     public void moveUpRepository(View view) {
-        Repository repository = getRepositoryFromView(view);
-        Repository targetRepository = repositoryDao.queryBuilder()
-                .where(RepositoryDao.Properties.Order.lt(repository.getOrder()))
-                .orderDesc(RepositoryDao.Properties.Order)
-                .limit(1).unique();
-        swapRepositoryOrder(repository, targetRepository);
+        repositoryFragment.moveUpRepository(view);
     }
 
     public void moveDownRepository(View view) {
-        Repository repository = getRepositoryFromView(view);
-        Repository targetRepository = repositoryDao.queryBuilder()
-                .where(RepositoryDao.Properties.Order.gt(repository.getOrder()))
-                .orderAsc(RepositoryDao.Properties.Order)
-                .limit(1).unique();
-        swapRepositoryOrder(repository, targetRepository);
-    }
-
-    private void swapRepositoryOrder(Repository repositoryA, Repository repositoryB) {
-        if (repositoryA == null || repositoryB == null)
-            return;
-
-        int order = repositoryA.getOrder();
-        repositoryA.setOrder(repositoryB.getOrder());
-        repositoryB.setOrder(order);
-        repositoryDao.updateInTx(repositoryA, repositoryB);
-        repositoryFragment.notifyRepositoriesChanged();
+        repositoryFragment.moveDownRepository(view);
     }
 
     public void popMoreMenu(View view) {
-        final Repository repository = getRepositoryFromView(view);
-        PopupMenu popupMenu = new PopupMenu(this, view);
-        popupMenu.inflate(R.menu.repository_more);
-
-        // Disable deleting if only one remaining.
-        long numRepositories = repositoryDao.queryBuilder()
-                .count();
-        if (numRepositories <= 1)
-            popupMenu.getMenu().findItem(R.id.menu_repository_delete).setEnabled(false);
-
-        popupMenu.setOnMenuItemClickListener(new PopupMenu.OnMenuItemClickListener() {
-            @Override
-            public boolean onMenuItemClick(MenuItem menuItem) {
-                return moreMenuItemClicked(menuItem.getItemId(), repository);
-            }
-        });
-        popupMenu.show();
+        repositoryFragment.popMoreMenu(view);
     }
 
-    private boolean moreMenuItemClicked(int itemId, Repository repository) {
-        if (itemId == R.id.menu_repository_sync) {
-            syncRepository(repository);
-        } else if (itemId == R.id.menu_repository_rename) {
-            renameRepository(repository);
-        } else if (itemId == R.id.menu_repository_delete) {
-            deleteRepositoryIfConfirmed(repository);
-        }
-        return true;
-    }
-
-    private void syncRepository(final Repository repository) {
-        View itemView =
-                (View) findViewById(R.id.repository_list).findViewWithTag(repository).getParent();
-        final ProgressBar progressBar =
-                (ProgressBar) itemView.findViewById(R.id.repository_progressbar);
-        new DownloadXmlAsyncTask(this, daoSession) {
-            @Override
-            protected void onPreExecute() {
-                super.onPreExecute();
-                progressBar.setProgress(0);
-                progressBar.setVisibility(View.VISIBLE);
-            }
-
-            @Override
-            protected void onProgressUpdate(Integer... values) {
-                super.onProgressUpdate(values);
-                progressBar.setProgress(values[0]);
-                progressBar.setVisibility(View.VISIBLE);
-            }
-
-            @Override
-            protected void onPostExecute(Integer result) {
-                super.onPostExecute(result);
-                progressBar.setVisibility(View.GONE);
-            }
-
-            @Override
-            protected void onCancelled(Integer result) {
-                super.onCancelled(result);
-                if (result != DownloadXmlAsyncTask.RESULT_CANCELLED)
-                    return;
-                progressBar.setVisibility(View.GONE);
-            }
-        }.execute(repository);
-    }
-
-    private void renameRepository(final Repository repository) {
-        final EditText editText = new EditText(this);
-        editText.setText(repository.getAlias());
-        editText.selectAll();
-
-        DialogInterface.OnClickListener onClickListener = new DialogInterface.OnClickListener() {
-            @Override
-            public void onClick(DialogInterface dialogInterface, int i) {
-                String newAlias = editText.getText().toString().trim();
-                if (newAlias.isEmpty())
-                    return;
-                repository.setAlias(newAlias);
-                repository.update();
-                repositoryFragment.notifyRepositoriesChanged();
-            }
-        };
-
-        new AlertDialog.Builder(this)
-                .setTitle(R.string.repository_rename_title)
-                .setCancelable(true)
-                .setView(editText)
-                .setNegativeButton(android.R.string.cancel, null)
-                .setPositiveButton(android.R.string.yes, onClickListener)
-                .show();
-        editText.requestFocus();
-    }
-
-    private void deleteRepositoryIfConfirmed(final Repository repository) {
-        // Generate message according to whether stars the on repository.
-        String message = String.format(getString(R.string.confirm_delete_repository_part1),
-                repository.getAlias());
-        List<Category> categories = repository.getCategories();
-        final List<Long> categoryIds = new ArrayList<Long>(categories.size());
-        for (Category category : categories)
-            categoryIds.add(category.getId());
-        long numStar = daoSession.getEntryDao().queryBuilder()
-                .where(EntryDao.Properties.CategoryId.in(categoryIds),
-                        EntryDao.Properties.Star.eq(true))
-                .count();
-        if (numStar > 0)
-            message += String.format(getString(R.string.confirm_delete_repository_part2),
-                    numStar);
-
-        // The actual code deleting the repository.
-        DialogInterface.OnClickListener onClickListener = new DialogInterface.OnClickListener() {
-            @Override
-            public void onClick(DialogInterface dialogInterface, int i) {
-                daoSession.runInTx(new Runnable() {
-                    @Override
-                    public void run() {
-                        daoSession.getEntryDao().queryBuilder()
-                                .where(EntryDao.Properties.CategoryId.in(categoryIds))
-                                .buildDelete().executeDeleteWithoutDetachingEntities();
-                        daoSession.getCategoryDao().queryBuilder()
-                                .where(CategoryDao.Properties.RepositoryId.eq(repository.getId()))
-                                .buildDelete().executeDeleteWithoutDetachingEntities();
-                        repository.delete();
-                    }
-                });
-                repositoryFragment.notifyRepositoriesChanged();
-            }
-        };
-
-        // Show alert dialog.
-        new AlertDialog.Builder(this)
-                .setCancelable(true)
-                .setTitle(R.string.confirm_delete_repository_title)
-                .setMessage(message)
-                .setNegativeButton(android.R.string.cancel, null)
-                .setPositiveButton(android.R.string.yes, onClickListener)
-                .show();
-    }
 }
