+++ app/src/main/java/org/sorz/lab/smallcloudemoji/adapters/RepositoryAdapter.java
@@ -42,12 +42,12 @@ public class RepositoryAdapter extends BaseAdapter {
 
     @Override
     public boolean areAllItemsEnabled() {
-        return false;
+        return true;
     }
 
     @Override
     public boolean isEnabled(int position) {
-        return false;
+        return true;
     }
 
     @Override
