+++ app/src/main/res/values-zh-rTW/strings.xml
@@ -51,4 +51,5 @@
     <string name="menu_repository_sync">同步</string>
     <string name="menu_repository_rename">重命名</string>
     <string name="repository_rename_title">新別名</string>
+    <string name="toast_repository_copied">鏈接已複製</string>
 </resources>
